+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.11.0 [future]
+
+### New Features
+BSplineCurve, Edge: simplify
+
+### Fixed
+- 2D conversion: create 2D function name in core_compiled
+- LineSegment, Arc, BSplineCurve: get_shared_section()
+
+### Changed
+- better surface3d plots
+
+
 ## v0.10.0 [Unreleased yet]
 
 ### New Features
@@ -95,6 +108,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * WiriMixin: from points: general method for Wire3D and 2D and for Contour2D and 3D. 
 * Added package.xml metadata in order to be listed in the FreeCAD Addon Manager 
 * Edge: local_discretization
+* ArcEllipse2d: point_at_abscissa, translation, split, point_distance.
 
 ### Fixed
 
@@ -117,7 +131,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * fix f string usage
 * Add some typings
 * Step: Step translator now handles some EDGE_LOOP inconsistencies coming from step files
-* Arc2d: point_belongs, abscissa. 
+* Arc2d: point_belongs, abscissa.
+* ArcEllipse2d: point_belongs, abscissa, init.
+
 
 ### Removed
 
@@ -157,7 +173,7 @@ is_inside_bbox, intersection_volume, distance_to_bbox, point_belongs, distance_t
 * VolumeModel: eq, volume, rotation, translation, frame_mapping, bounding_box, plot
 * Wire: extract_with_points, split_with_two_points
 * Arc2d: point_belongs, abscissa.
-
+* ArcEllipse2d: point_belongs, abscissa, init, translation, split, point_at_abscissa, point_distance.
 ### CI
 - add spell check to pylint with pyenchant
 - make code_pydocstyle more explicit